Deprecated require 'valvat/local' (BREAKING)
